# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Serverless Harness control plane
  version: 1.0.0
  description: |
    Multi-user control plane: authenticated API, owned sessions, per-user credentials (MU1).
    This document covers the CONTROL PLANE's routes only. The data plane also serves `/v1` paths
    -- `/v1/turn` and `/v1/runs`, on the harness Knative Service -- and they are deliberately not
    described here; "the `/v1` API contract" in spec §1/§4 is therefore wider than this file.
    `/internal/credentials` is the data-plane-only credential exchange and is not part of the
    public surface; it is documented here because it is part of the contract between the tiers.
servers:
  - url: http://sh-control-plane.default.svc:8080
    description: in-cluster
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ed25519 (`alg: EdDSA`) JWT minted by the control plane. Two shapes: an API token
        (`scope: ["api"]`, 1 hour) for `/v1` routes, and a session token (`scope: ["turn:write"]`
        plus `sid`, 5 minutes) for the data plane's `/turn`. No credential travels in the token.
    exchangeToken:
      type: http
      scheme: bearer
      description: The shared `SH_EXCHANGE_TOKEN`, mounted into both tiers from one Secret.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          # A real enum, NOT a prose list: this is the half a client generator and a schema validator
          # consume, and the half openapi-contract.test.ts compares against CP_ERROR_CODES. Keep it
          # sorted the same way as the source array for a readable diff when a code is added.
          enum:
            - invalid_json
            - invalid_request
            - token_required
            - token_invalid
            - token_expired
            - subject_conflict
            - unauthorized
            - forbidden
            - session_not_found
            - session_mismatch
            - credential_required
            - credential_ambiguous
            - credential_not_found
            - credential_unavailable
            - endpoint_unresolved
            - authorization_pending
            - redis_unavailable
            - internal_error
          description: snake_case error code.
        message:
          type: string
          description: |
            Human-readable detail. ABSENT on `internal_error` by design -- an arbitrary error's text
            can carry a Redis connection string or a presented token, and this body reaches an
            arbitrary caller (errors.ts).
        sessionId: { type: string }
...
  responses:
    Unauthorized:
      description: no token, a bad token, or a failed exchange authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /v1/auth/device:
    post:
      operationId: startDeviceAuth
      summary: Start a GitHub OAuth device-flow login
      description: |
        GitHub's user-login flow is plain OAuth 2.0, not OIDC (spec §5.1). The device flow is what a
        shell script can complete, and it needs no client secret. The GitHub OAuth app must have
        device flow ENABLED -- it is off by default (spec §5.1.1).
      security: []
      responses:
        '200':
          description: the code the operator types at verificationUri
          content:
            application/json:
              schema:
                type: object
                required: [deviceCode, userCode, verificationUri, interval, expiresIn]
                properties:
                  deviceCode: { type: string }
                  userCode: { type: string }
                  verificationUri: { type: string }
                  interval: { type: integer }
                  expiresIn: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
...